Better implementation of collections
Now every element passed to a collection is converted to a path, if it
isn't already a path. We are also inferring types from the elements
passed to `Collection.create`.

Both of these things allow us to pass single paths to the core function
(e.g. within a `.forEach()` callback) and get a properly typed
collection back.