Add runner
This adds a simple runner with which the transform can be executed.
Installing the package globally adds the executable jscs. Example usage:

jscs -t transformFile files to transform
jsnc files to transform

If no transform file is specified, the runner will look for the file
transform.js in the current working directory.

The file should be a node module that exports a function, taking the
source of a file and returning the transformed source.
Inside the module, the variable jscs refers to tool.

Example file:

module.exports = function(source) {
  return jscs(source).findVariableDeclarators('foo').renameTo('xyz').toSource();
}

For now we only provide a simple report on the console. At some point we
should add more reporters, e.g. for machine readable reports.