Update runner
Still hacky, but a bit more structured now. New options

-d, --dry: Dry run. Process files but don't write them
-c, --cpus: The number of cores to use
-v, -vv: Verbosity level. Only shows errors by default. -v shows
skipped and unchanged files. -vv shows transformed files

New features:

If the transform function returns a falsy value, the file is marked as
to be "skipped". For example if the condition for the transformation is
not fulfilled.

Transform files can pass a name and a counter to the global `stat`
function. This can be used to collect information about the files.
Stats will only be recorded when running in dry mode.