Improve description of -v option
I found the description for this option to be confusing, partly because
I don't think "repeate" is actually a word, and partly because if it
was a typo for "repeat", it hardly makes sense to me here. In any case,
these words seem unnecessary and can probably just be removed entirely.