Improve formatting of Extensibility readme section
There was an intentional line break here that was intended to help
people parse the two types of extensions. However, it is pretty unusual
to see a line break in the middle of a paragraph like that, and this
detail was accidentally lost when I recently removed some trailing
whitespace. Replacing this formatting with some bold words accomplishes
the same goal but in a more expected manner.