Add source code - AST Node interpolation
This adds three tagged templates to the core API:

- jscodeshift.template.expression
- jscodeshift.template.statement
- jscodeshift.template.statements

Their purpose is to make it easier to replace nodes with more complex
constructs, without having the build the AST yourself.
For example, the following replaces all matching identifiers with a function
call:

```
let {expression} = jscodeshift.template;

// replaces all `bar` identifiers with the function call `foo(bar)`
jscodeshift(src)
  .find('Identifier', {name: 'bar'})
  .replaceWith(path => expression`foo(${path.node})`);
```

The difference between the three function is just how the provided
code snippet is parsed:

- `expression` returns an expression node
- `statement` returns a statement node
- `statements` returns an array of statement nodes

It is now also possible to replace a node with multiple nodes. This is
especially useful for replacing a single statement with multiple
statements (still have to experiment with expression -> multiple
expressions replacements).