Fix printing (#28)
There have been multiple reports about the printed source missing
trailing line breaks. The reason for the line break missing is that this
information is contained in the `File` node, which is returned by
`recast.parse`.
However, because we build a node collection from the `Program` node
instead of the `File` node, we are loosing this information.
Using the `File` node fixes the issue and doesn't seem to break any
existing functionality.