Add cli generator helper for codemod collections
This change makes it easier to create codemod collections with a custom
cli. It will make it possible to publish a collection with jscodeshift
as a dependency. Furthermore, it will not make it necessary for the user
to know the structure of the collection or to have it cloned and refer
to a specific path.

Example usage for a bin file:

```
const resolve = require('path').resolve;
const cli = require('jscodeshift').cli;

const pkg = require('./package.json');

cli({
  name: 'codemods',
  packageVersion: pkg.info,
  options: { quote: 'single' },
  resolve(name) {
    return resolve(__dirname, `./transforms/${name}.js`);
  },
});

```