Fix `renameTo` not taking property shorthands into account
This fixes the `renameTo` helper method not properly renaming
object properties that are shorthands.

The previous version of the code would rename the property's
`value` correctly, but since `shorthand` was still set as `true`,
the output result would still be the old variable as a shorthand
property.

Now, if an identifier to be renamed is of type `Property` and
is a shorthand (but not a method), we flip the `shorthand` value
to be `false`.