Add simple reporting API
This diff adds a new method to the API passed to transformers, `report`.
This function accepts a string and will print that string (with the file
path) to stdout.

This makes it possible for transformers to report arbitrary information
to other tools that process jscodeshift's output.

In the future the API can be extended to be able to consume the reports
in a programmatic way.