fix: preserve file extension for remote file
There is behavior in this program that depends on the extension
of the transformFile, notably how babel runs (or doesn't).

There was a bug when requesting a remote transformFile authored as
an ES6 module because of the temp filename being used.

This commit preserves the original file extension used when writing
the remote file to the temp one.